moved save high school to admin
