+++ high_school/templates/high_school/save_high_schools.html
@@ -1,9 +1,14 @@
-{% extends 'high_school/index.html' %}
+{% extends 'admin/base_site.html' %}
 
-{% block save_highschool %}
-    <form action="{% url 'dashboard:high_school:save'  %}" method="POST">
+{% block content %}
+    <form id="save_hs_form" action="{% url 'admin:save_high_schools' %}" method="POST">
         {% csrf_token %}
         {{ form }}
         <input type="submit" value="Save"/>
     </form>
+    {% if errors %}
+        <span>{{ errors }}</span>
+    {% elif response %}
+        <span>{{ response }}</span>
+    {% endif %}
 {% endblock %}
